A few small bits of cleanup
