+++ pkg/abstract/primitive-maps.go
@@ -14,15 +14,23 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package util
+package abstract
 
-// SliceEqual compares two slices, element positions considered significant
-func SliceEqual[Elt comparable](slice1, slice2 []Elt) bool {
-	if len(slice1) != len(slice2) {
+// PrimitiveMapGet exposes the indexing functionality of a primitive map as a func
+func PrimitiveMapGet[Key comparable, Val any](rep map[Key]Val) func(Key) (Val, bool) {
+	return func(key Key) (Val, bool) {
+		val, has := rep[key]
+		return val, has
+	}
+}
+
+func PrimitiveMapEqual[Key, Val comparable](map1, map2 map[Key]Val) bool {
+	if len(map1) != len(map2) {
 		return false
 	}
-	for idx, elt1 := range slice1 {
-		if slice2[idx] != elt1 {
+	for key, val1 := range map1 {
+		val2, have2 := map2[key]
+		if !(have2 && val1 == val2) {
 			return false
 		}
 	}
